Spark Migration (#2)
* setup databricks cli connection

* added databricks env vars

* set up run sql workflow

* Created DatabricksSQLOperator and run_custom_spark_job function

* Added create table script

* Added json_args to SparkScriptOperator: now can dump any size params to read into python script

* create sql and create_table operators for spark

* set up create tables pipeline

* Updated bug in prod dump table

* Built cj download and product info processing pipelines

* BIg Commit: Refactor of spark script operator, got product emb script working and deployed

* Similar items pipeline complete

* SIMILAR ITEMS PIPELINE DEPLOYED!

* Fixed bottlenecks in product_ML and similar items, changed similar items write to postgres to staging, various other things

* Updated postgres index definitions

* User events stream set up

* added product stats pipeline

* user recs pipeline deployed!

+++ src/airflow_tools/queries/postgre_queries.py
@@ -1,6 +1,6 @@
 from src.defs.postgre import utils
 
-def create_table_query(table_name: str, columns: list,
+def create_table_query(table_name: str, columns: list, is_prod: bool = False,
                        tail: str="", drop: bool=False):
     query = "BEGIN TRANSACTION;\n"
     if drop == True:
@@ -12,10 +12,13 @@ def create_table_query(table_name: str, columns: list,
             query += col
         else:
             query += f"\t{col['name']} {col['type']} {col['mode']}"
+            if is_prod:
+                query += " " + col.get("prod", "")
         if i != len(columns) - 1:
             query += ","
         query += "\n"
 
+    tail = tail if is_prod else ""
     query += f") {tail};\n"
     query += "END TRANSACTION;"
     return query
@@ -35,7 +38,8 @@ def staging_to_live_query(table_name: str,
                           staging_name: str, mode: str,
                           tail: str = "",
                           key: str = None,
-                          columns: list = []):
+                          columns: list = [],
+                          drop_staging: bool = False):
     query = "BEGIN TRANSACTION;\n"
     if mode == "WRITE_TRUNCATE":
         query += write_truncate(table_name, staging_name, columns)
@@ -44,7 +48,7 @@ def staging_to_live_query(table_name: str,
     if mode == "UPSERT":
         query += upsert(table_name, staging_name, key, columns)
     query += tail
-    query += f"DROP TABLE IF EXISTS {staging_name};\n"
+    query += f"DROP TABLE IF EXISTS {staging_name};\n" if drop_staging else ""
     query += "END TRANSACTION;"
     return query
 
@@ -86,14 +90,19 @@ def upsert(table_name, staging_name, key, columns):
 
 def export_rows(table_name,
     export_table_name,
-    columns="*",
+    columns,
     FILTER="",
-    delete=False):
+    delete=False,
+    clear_export_table=False):
 
     SQL = f"""
     BEGIN TRANSACTION;
+    """
+    SQL += f"DELETE FROM {export_table_name};" if clear_export_table else ""
+
+    SQL += f"""
     INSERT INTO {export_table_name}
-    SELECT * FROM {table_name}
+    SELECT {columns} FROM {table_name}
     {FILTER};
     """
     if delete:
@@ -102,7 +111,9 @@ def export_rows(table_name,
         {FILTER};
         END TRANSACTION;
         """
-    return SQL 
+    else:
+        SQL +=" END TRANSACTION;"
+    return SQL
 
 def _build_distinct_filter(columns):
     if len(columns) == 0:
